unit cards in pdf cleaned up and mostly match the unit card preview unit card display pulled into its own class
